+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased changes
+- Added support for Transfer With Schedule.
+- Added support for Transfer With Schedule With Memo.
 - Added support for Transfer to public wallet.
 - Added support for Transfer To Encrypted wallet.
 - Added support for Encrypted Transfer.